requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
httpx==0.25.2
python-dotenv==1.0.0
openai==1.3.7
